+++ LeMaitreBusiness/src/br/cefetmg/LeMaitre/model/dao/TableDAO.java
@@ -13,9 +13,9 @@ import br.cefetmg.LeMaitre.model.exception.PersistenceException;
  * @author Thalesgsn
  */
 public interface TableDAO {
-    public Integer insert(Table table) throws PersistenceException;
+    public Long insert(Table table) throws PersistenceException;
     public boolean update(Table table) throws PersistenceException;
-    public boolean remove(Integer tableID) throws PersistenceException;
-    public Table getTableByID(Integer tableID) throws PersistenceException;
-    public boolean containsThisTableID(Integer tableID) throws PersistenceException;
+    public boolean remove(Long tableID) throws PersistenceException;
+    public Table getTableByID(Long tableID) throws PersistenceException;
+    public boolean containsThisTableID(Long tableID) throws PersistenceException;
 }
